chore: use preexisting http service, expand config client for assets and chains
